table with popup table
